backend/requirements.txt
# Foundit Backend Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# AI & Machine Learning
sentence-transformers>=2.3.1
torch>=2.0.0
transformers>=4.36.2

# Vector Database
faiss-cpu>=1.7.4

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# LLM Providers
# For local LLM (Ollama) - only needs requests
requests>=2.31.0

# For OpenAI (optional - uncomment if needed)
# openai==1.12.0

# For Anthropic Claude (optional - uncomment if needed)
# anthropic==0.18.0

# Utilities
numpy>=1.26.3

# OAuth & Authentication
authlib==1.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
